+++ tests/commands-git/branch.test.ts
@@ -140,6 +140,7 @@ describe('BranchOptions', () => {
     [{ '--no-merged': 'c233c7a' }, ['--no-merged', 'c233c7a']],
     [{ '--no-merged': ['c233c7a'] }, ['--no-merged', 'c233c7a']],
     [{ '--no-merged': ['c233c7a', 'ab2c420'] }, ['--no-merged', 'c233c7a', '--no-merged', 'ab2c420']],
+    [{ '--omit-empty': null }, ['--omit-empty']],
     [{ '--points-at': null }, ['--points-at']],
     [{ '--points-at': 'HEAD' }, ['--points-at', 'HEAD']],
     [{ '--remotes': null }, ['--remotes']],